Refactoring to consolidate the two methods of defining acquisition functions. Modules now call fucntions from the  directory where appropriate. Also some renaming of subfolders for consistency, and updates to imports to attempt obvious failure from import issues.
