--- gpytorch/gpytorchopt/acquisition_functions_functional/__init__.py
@@ -1,1 +1,0 @@
-#!/usr/bin/env python3
