+++ gpytorch/gpytorchopt/bayesian_optimization_example3_MCMC.py
@@ -1,15 +1,15 @@
 import torch
 import gpytorch
 from gpytorchopt.additive.additive_structure_gp_model import AdditiveStructureGPModel
-from gpytorchopt.acquisition_function_strategies.discrete_set import (
+from gpytorchopt.acquisition.strategy.discrete_set import (
     EnsembleDiscreteSetEvaluationStrategy,
     DiscreteSetRandomStrategy,
 )
-from gpytorchopt.additive.acquisition_function_strategies import AdditiveStrategy, DimScanAdditiveStrategy
+from gpytorchopt.additive.strategy import DimScanAdditiveStrategy
 from gpytorchopt import BayesianOptimization
 from gpytorchopt.additive.structure_discovery import MetropolisHastingAdditiveStructureSelector
 import ghalton
-from gpytorchopt.acquisition_functions import ExpectedImprovement
+from gpytorchopt.acquisition import ExpectedImprovement
 
 # Define the function
 func1 = False
@@ -105,7 +105,7 @@ class BayesianOptimizationModelMCMC(BayesianOptimization):
         return true_candidate, function_value
 
 
-initial_kernel = gpytorch.kernels.AdditiveKernel(gpytorch.kernels.RBFKernel(active_dims=list(range(n_dims))))
+initial_kernel = gpytorch.kernels.kernel.AdditiveKernel(gpytorch.kernels.RBFKernel(active_dims=list(range(n_dims))))
 likelihood = gpytorch.likelihoods.GaussianLikelihood(log_noise_bounds=(-5, 1))  # .cuda()
 initial_model = AdditiveStructureGPModel(
     torch.Tensor([0]), torch.Tensor(), likelihood, initial_kernel
