+++ gpytorch/gpytorchopt/gpytorchopt/acquisition/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+
+from .modules import AcquisitionFunction
+from .modules import ExpectedImprovement
+from .modules import MaxValueEntropySearch
+from .modules import ProbabilityOfImprovement
+from .modules import UpperConfidenceBound
