+++ gpytorch/gpytorchopt/gpytorchopt/acquisition/functional/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+
+from .acquisition import (
+    expected_improvement,
+    max_value_entropy_search,
+    posterior_mean,
+    probability_of_improvement,
+    upper_confidence_bound,
+)
