+++ gpytorch/gpytorchopt/gpytorchopt/acquisition/functional/knowledge_gradient.py
