--- gpytorch/gpytorchopt/gpytorchopt/acquisition_functions/__init__.py
@@ -1,5 +1,0 @@
-from .acquisition_function import AcquisitionFunction
-from .expected_improvement import ExpectedImprovement
-from .max_value_entropy_search import MaxValueEntropySearch
-
-__all__ = [AcquisitionFunction, ExpectedImprovement, MaxValueEntropySearch]
