+++ gpytorch/gpytorchopt/gpytorchopt/additive/additive_structure_gp_model.py
@@ -1,5 +1,5 @@
 import gpytorch
-from gpytorch.kernels import AdditiveKernel
+from gpytorch.kernels.kernel import AdditiveKernel
 import torch
 from torch import nn
 import copy
